Add entry, restore, and dump system
The basic entry system is based on reactions. Some channels and roles
are hardcoded due to this not being scaled and that I'm lazy.

If the bot goes down, the db is restored and the giveaway is restarted
from the previous point.

1. If the current time passes the giveaway lock timestamp, immediately
lock the giveaway upon bot restart
2. If the current time passes the giveaway winner announcement, immediately
anounce the winner upon bot restart

These events are fired from cache_ready since it requires a message
from the cache.

The entry dump system is made to log who enters/leaves the giveaway
in a minimal format. It mentions the user and sees if he/she has enters/leaves
the giveaway.

If the bot goes down and the giveaway is still live, the user can DM
the giveaway setup person (aka me) and I can manually add an entry.

Signed-off-by: kingbri <bdashore3@gmail.com>